+++ arch/core/decoder/decoder.scala
@@ -8,25 +8,20 @@ import chisel3._
 class Decoder(implicit p: Parameters) extends Module {
   override def desiredName: String = s"${p(ISA)}_decoder"
 
-  val table    = DecodeTableFactory.get(p(ISA)) match {
-    case Some(t) => t
-    case None    => throw new Exception(s"DecodeTable for ISA ${p(ISA)} not found!")
-  }
-  val sigsType = DecodeCtrlSigsFactory.get(p(ISA)) match {
-    case Some(s) => s
-    case None    => throw new Exception(s"DecodeCtrlSigs for ISA ${p(ISA)} not found!")
+  val utils = DecoderUtilitiesFactory.get(p(ISA)) match {
+    case Some(u) => u
+    case None    => throw new Exception(s"Decoder utilities for ISA ${p(ISA)} not found!")
   }
 
   val instr   = IO(Input(UInt(p(ILen).W)))
-  val decoded = IO(Output(sigsType.createBundle()))
+  val decoded = IO(Output(utils.createBundle()))
 
-  decoded := sigsType.decode(instr, table.table)
+  decoded := utils.decode(instr, utils.table)
 }
 
 // Test
 object DecoderInit {
-  val rv32iDecodeTable = new RV32IDecodeTable
-  val rv32iCtrlSigs    = RV32ICtrlSigsDecoder
+  val rv32iUtils = RV32IDecoderUtilities
 }
 
 object DecoderTest extends App {
